Fixes #2604: "Refactor: Decompose StateTracker 138 methods into dom... (#2650)
* Fixes #2604: Decompose StateTracker into domain-based mixin modules

Split the monolithic StateTracker (1,308 lines, 138 methods) into 9
domain-based mixins inside a `src/state/` package while preserving the
public API (`from state import StateTracker`). The facade class inherits
all mixins via multiple inheritance.

Mixins: IssueState, WorktreeState, HITLState, ReviewState, EpicState,
LifetimeStats, SessionState, WorkerState, ReportState.


* review: add missing tests for _maybe_migrate_worker_states (PR #2650)

The worker-state migration path (_maybe_migrate_worker_states) ran on every
StateTracker load but had no test coverage. Added two tests to
test_state_mixin_decomposition.py:
- verify legacy bg_worker_states are migrated into worker_heartbeats on load
- verify migration is skipped when worker_heartbeats are already populated


* review: add missing IssueStateMixin tests (PR #2650)

Six untested public methods in IssueStateMixin had no coverage:
mark_pr, set/get_worker_result_meta, record_outcome (including the
previous-outcome counter-decrement branch), get_all_outcomes,
record_hook_failure (including the 500-entry cap), and get_hook_failures.


---------
